+++ CAREN/src/main/java/com/booboo/CAREN/Parser/Nearby.java
@@ -21,7 +21,7 @@ public class Nearby implements Node{
     }
 
     @Override
-    public double evaluate() throws SyntaxError {
+    public int evaluate() throws SyntaxError {
         int min = Integer.MAX_VALUE;
         for(Virus v : listVirus){
             int newmin = findLocation(host.getPos().PosX(), host.getPos().PosY(), v.getPos().PosX(), v.getPos().PosY());
